Revert "Ait 13 update readme md"
