--- .flake8
@@ -1,4 +1,0 @@
-[flake8]
-max-line-length = 100
-ignore = W503
-exclude = .git,__pycache__,.venv
