--- .github/workflows/style.yml
@@ -1,25 +1,0 @@
-name: Python style check
-
-on: [push, pull_request]
-
-jobs:
-  style:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Setup Python 3.8
-        uses: actions/setup-python@v5
-        with:
-          python-version: 3.x
-
-      - name: Install Dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install flake8
-
-      - name: Run flake8
-        run: flake8 .
-        working-directory: .
-        continue-on-error: false
