--- bot/commands/start.py
@@ -1,9 +1,0 @@
-from aiogram import Router, types
-from aiogram.filters import Command
-
-router = Router()
-
-
-@router.message(Command("start"))
-async def start_command(message: types.Message):
-    await message.answer("Hi! Send me a voice message to get started.")
